refactor(naming): finish Dashboard → Overview rename
Page was renamed earlier but the symbol, components/dashboard dir, and
i18n namespace still carried the old name.
