+++ frontend/src/i18n/locales/en/overview.json
@@ -1,5 +1,5 @@
 {
-  "title": "Dashboard",
+  "title": "Overview",
   "planCta": {
     "title": "Plan {{month}}",
     "description": "Carries last month's actuals into this month's budgets.",
