+++ frontend/src/pages/Income.tsx
@@ -23,7 +23,7 @@ import { getCurrentFinancialMonth, getFinancialMonthRange, getPreviousFinancialM
 import { fetchHistoryByKey } from "@/utils/carryForward";
 import { computeSmartDefault } from "@/services/smartDefaults";
 import { reportSuccess, reportFailure, isDown } from "@/utils/outageMonitor";
-import { useMonthlyReviewStatus } from "@/components/dashboard/MonthlyReviewWizard";
+import { useMonthlyReviewStatus } from "@/components/overview/MonthlyReviewWizard";
 
 import { IncomePageSkeleton } from "@/components/shared/skeletons/PageSkeletons";
 import { Skeleton } from "@/components/ui/skeleton";
@@ -465,7 +465,7 @@ const Income = () => {
           <AlertContent>
             <AlertTitle>This month's review hasn't been finalized.</AlertTitle>
             <AlertDescription>
-              Edits are locked until the Monthly Review is complete. Use the wizard on the Dashboard to review and finalize.
+              Edits are locked until the Monthly Review is complete. Use the wizard on the Overview to review and finalize.
             </AlertDescription>
           </AlertContent>
           <Button asChild size="sm" variant="outline" className="shrink-0">
